requirements.txt
tensorflow-cpu == 2.12.0
seaborn
umap
tensorflow-hub
tensorflow-text == 2.12.0